Finally its stable enough to merge here

+++ utils/vrcAPI.ts
@@ -1,11 +1,11 @@
 import * as vrchat from "vrchat";
-import { LogManager } from './logger';
-import { env } from "bun";
+import {LogManager} from './logger';
+import {env} from "bun";
 import * as fs from "fs";
 import axios from "axios";
 import { Cookie, CookieJar } from 'tough-cookie';
 import type { AxiosResponse } from "axios";
-import type { VRSpaceVRCUserAvatar } from "../interfaces/apiHelper";
+import type {VRChatCookieFormat, VRSpaceVRCUserAvatar} from "../interfaces/apiHelper";
 
 // I don't fucking know what I'm doing, but it makes 'axios.default.jar' happy, and that's all it matters.
 declare module 'axios' {
@@ -15,7 +15,7 @@ declare module 'axios' {
 }
 
 // Cookie Jar Stuff
-let cookies: string; 
+let cookies: string;
 
 // Logger Stuff
 const debugType: string = 'error';
@@ -23,8 +23,8 @@ const logger: LogManager = new LogManager(debugType, 'VRCHAT-API');
 
 if(await Bun.file('./.env').exists()) {
     if((env.VRC_USERNAME === "" || env.VRC_PASSWORD === "") ||
-    (!env.VRC_USERNAME || !env.VRC_PASSWORD) ||
-    (env.VRC_USERNAME === "your_vrchat_username" || env.VRC_PASSWORD === "your_vrchat_password")) {
+        (!env.VRC_USERNAME || !env.VRC_PASSWORD) ||
+        (env.VRC_USERNAME === "your_vrchat_username" || env.VRC_PASSWORD === "your_vrchat_password")) {
         logger.warn("Please set your VRC_USERNAME and VRC_PASSWORD in your .env file")
         process.exit(1);
     }
@@ -34,7 +34,7 @@ else {
     process.exit(1);
 }
 
-
+logger.debug("Running Axios version: "+axios.VERSION)
 
 // If the cookies file exists, load it
 // Otherwise, throw an error
@@ -75,13 +75,12 @@ const configuration: vrchat.Configuration = new vrchat.Configuration({
     }
 });
 
-// API Methods 
+// API Methods
 const AuthenticationApi: vrchat.AuthenticationApi = new vrchat.AuthenticationApi(configuration);
 const FriendsApi: vrchat.FriendsApi = new vrchat.FriendsApi(configuration);
 const UsersApi: vrchat.UsersApi = new vrchat.UsersApi(configuration);
 const NotificationsApi: vrchat.NotificationsApi = new vrchat.NotificationsApi(configuration);
 const InstancesApi: vrchat.InstancesApi = new vrchat.InstancesApi(configuration);
-const AvatarApi: vrchat.AvatarsApi = new vrchat.AvatarsApi(configuration);
 
 logger.success("VRC API is configured")
 
@@ -100,31 +99,13 @@ This token is also set up by the [Set-Cookie] header, so you need to save it in 
 For Auto-Login, you can save the cookie jar in a file, and load it in the Axios defaults for the API.
 (See the function 'setAuthCookie')
 */
-async function doLogin(forceLogin?: boolean, onlySaveAuthCookie?: boolean): Promise<vrchat.CurrentUser | undefined>{
-    logger.working("Executing VRChat login...")
-    try {
-        const resp: AxiosResponse<vrchat.CurrentUser> = await AuthenticationApi.getCurrentUser();
-        const currentUser: vrchat.CurrentUser = resp.data;
-        if(onlySaveAuthCookie) { 
-            setAuthCookie();
-            return;
-        }
-        if(forceLogin) {
-            logger.info("Forcing 2FA Login and save cookies")
-            deleteCookieFile();
-            const twoFactorCode: string | null = prompt("[*] Please enter your two factor code: ")?.toString() ?? "";
-            const verifyResp: AxiosResponse<vrchat.Verify2FAResult> = await AuthenticationApi.verify2FA({ code: twoFactorCode });
-            if (verifyResp.data.verified) {
-                logger.success("Verified Successfully, welcome to VRSpace!");
-                setAuthCookie();
-                return;
-            }
-        }
-        if (!currentUser.displayName) {
-            const twoFactorCode: string | null = prompt("[*] Please enter your two factor code: ")?.toString() ?? "";
-            logger.debug(`Two factor code: ${twoFactorCode}`);
-
-            const verifyResp = await AuthenticationApi.verify2FA({ code: twoFactorCode });
+async function doLogin(forceLogin?: boolean): Promise<vrchat.CurrentUser>{
+    return AuthenticationApi.getCurrentUser().then(async (resp) => {
+        if (forceLogin || !resp.data.displayName) {
+            if (forceLogin) deleteCookieFile();
+            const verifyResp = await AuthenticationApi.verify2FA({
+                code: prompt("[*] Please enter your two factor code: ")?.toString() ?? ""
+            });
             if (verifyResp.data.verified) {
                 logger.success("Verified Successfully, welcome to VRSpace!");
                 setAuthCookie();
@@ -133,38 +114,33 @@ async function doLogin(forceLogin?: boolean, onlySaveAuthCookie?: boolean): Prom
         const myself = await getMyself();
         logger.success("Welcome Back "+myself?.displayName+" to VRSpace!");
         return resp.data;
-    } catch (e) {
+    }).catch((e) => {
         const errorResponse = e as any;
         if (errorResponse.response && errorResponse.response.status) {
             switch (errorResponse.response.status) {
                 case 400:
-                    console.log("[✘] Token is invalid");
-                    break;
+                    throw new Error("Token is invalid");
                 case 403:
-                    console.log("[✘] Two factor authentication is required");
-                    break;
+                    throw new Error("Two factor authentication is required")
                 default:
-                    console.log("[✘] Unknown error, please see the following for more information:");
-                    console.log(errorResponse.response.data);
-                    break;
+                    throw new Error("Unknown error, please see the following for more information:" + errorResponse.response.data);
             }
-        } else {
-            console.log("Unknown error");
         }
-    }
+        throw errorResponse;
+    });
 }
 
-
-async function loginAndSaveCookies(onylWS?: boolean): Promise<void> {
-    if(onylWS) {
-        await doLogin(false, true);
-    } else {
-        await doLogin(true);
-    }
-    
+async function loginAndSaveCookies(): Promise<void> {
+    await doLogin(true);
 }
+
 
 // Saves the AuthCookie and TwoFactorAuth from the CookieJar into a JSON file
+/**
+ * Set the authentication cookie for API calls.
+ * @param {string} authCookie - The authentication cookie value.
+ * @return {void}
+ */
 function setAuthCookie(authCookie?: string): void {
     const jar: CookieJar | undefined = (axios.defaults)?.jar;
     if(!jar) {
@@ -176,12 +152,12 @@ function setAuthCookie(authCookie?: string): void {
         jar.setCookie(
             new Cookie({key: 'auth', value: authCookie}),
             'https://api.vrchat.cloud'
-        );
+        ).then(() => logger.write("authCookie written to Cookie JSON file"));
     }
     jar.setCookie(
         new Cookie({key: 'apiKey', value: 'JlE5Jldo5Jibnk5O5hTx6XVqsJu4WJ26'}),
         'https://api.vrchat.cloud'
-    );
+    ).then(() => logger.write("apiKey written to Cookie JSON file"));
     fs.writeFileSync("./cookies.json", JSON.stringify(jar.toJSON()));
     logger.success("Cookies file saved")
 }
@@ -206,21 +182,19 @@ function deleteCookieFile(): void {
  *
  * @returns {Promise<vrchat.LimitedUser[] | undefined>} A promise that resolves to an array of online friends. If an error occurs, the promise will resolve to undefined.
  */
-async function seeOnlineFriends(): Promise<vrchat.LimitedUser[] | undefined> {
+async function seeOnlineFriends(): Promise<vrchat.LimitedUser[]> {
     logger.write("Getting online friends")
-    try {
-        const resp: AxiosResponse<vrchat.LimitedUser[]> = await FriendsApi.getFriends();
+    return FriendsApi.getFriends().then((resp) => {
         return resp.data;
-    } catch (e) {
+    }).catch(async (e) => {
         if(axios.isAxiosError(e)) {
             if(e.response?.status === 401) {
                 console.log("[✘] Token maybe invalid");
                 await doLogin(true);
             }
         }
-        else
-            console.error(e);
-    }
+        throw e;
+    });
 }
 
 
@@ -231,14 +205,10 @@ async function seeOnlineFriends(): Promise<vrchat.LimitedUser[] | undefined> {
  * @returns {Promise<vrchat.User | undefined>} - A Promise that resolves with the user information if found, otherwise undefined.
  */
 async function getUserInfo(userId: string): Promise<vrchat.User | undefined>{
-    logger.write("Getting user info for: "+userId)
-    try {
-        const resp: AxiosResponse<vrchat.User> = await UsersApi.getUser(userId)
+    logger.working("Getting user info for: "+userId)
+    return UsersApi.getUser(userId).then((resp) => {
         return resp.data;
-    } catch (e: any) {
-        console.error(e);
-        console.log(e.response?.data)
-    }
+    })
 }
 
 /**
@@ -247,15 +217,11 @@ async function getUserInfo(userId: string): Promise<vrchat.User | undefined>{
  * @param {string} username - The username to search for.
  * @return {Promise<vrchat.LimitedUser[] | undefined>} - A promise that resolves with an array of LimitedUser objects if the search is successful, or undefined if the search fails.
  */
-async function searchUser(username: string): Promise<vrchat.LimitedUser[] | undefined>{
+async function searchUser(username: string): Promise<vrchat.LimitedUser[]>{
     logger.write("Searching for user: "+username)
-    try {
-        const resp: AxiosResponse<vrchat.LimitedUser[]> = await UsersApi.searchUsers(username)
-        if(resp)
-            return resp.data;
-    } catch (e) {
-        console.error(e);
-    }
+    return UsersApi.searchUsers(username).then((resp) => {
+        return resp.data;
+    });
 }
 
 
@@ -264,13 +230,10 @@ async function searchUser(username: string): Promise<vrchat.LimitedUser[] | unde
  *
  * @returns {Promise<vrchat.Notification[] | [] | undefined>} A promise that resolves to an array of notifications, or an empty array, or undefined if an error occurs.
  */
-async function getNotifications(): Promise<vrchat.Notification[] | [] | undefined> {
-    try {
-        const resp: AxiosResponse<vrchat.Notification[]> = await NotificationsApi.getNotifications();
+async function getNotifications(): Promise<vrchat.Notification[] | []> {
+    return NotificationsApi.getNotifications().then((resp) => {
         return resp.data;
-    } catch (e) {
-        console.error(e);
-    }
+    })
 }
 
 /**
@@ -282,14 +245,11 @@ async function getNotifications(): Promise<vrchat.Notification[] | [] | undefine
  * @throws {Error} If an error occurs during the logout operation
  */
 async function doLogout(deleteCookies?: boolean): Promise<void> {
-    try {
-        const resp: AxiosResponse<vrchat.Success> = await AuthenticationApi.logout();
+    return AuthenticationApi.logout().then((resp) => {
         logger.info(resp.data);
         if(deleteCookies)
             deleteCookieFile();
-    } catch (e) {
-        console.error(e);
-    }
+    })
 }
 
 
@@ -311,73 +271,102 @@ async function getAuthCookie(): Promise<string> {
             throw new Error("No auth cookie found, please login first");
         }
     })/*.catch((e) => {
-        logger.fatal("There was a error while reading the cookie file: ");
+        logger.fatal("There was an error while reading the cookie file: ");
         logger.error(e);
     })*/;
 }
 
 
-async function getInstanceInfo(worldId: string, instanceId: string): Promise<vrchat.Instance | undefined> {
+/**
+ * Retrieves information about a specific instance.
+ *
+ * @param {string} worldId - The ID of the world in which the instance is located.
+ * @param {string} instanceId - The ID of the specific instance.
+ * @returns {Promise<vrchat.Instance>} A Promise that resolves to an object containing instance data.
+ * @throws {Error} If there is an error while trying to retrieve the instance data.
+ */
+async function getInstanceInfo(worldId: string, instanceId: string): Promise<vrchat.Instance> {
     logger.write("Getting instance data for: "+worldId+" - "+instanceId)
-    try {
-        const resp = await InstancesApi.getInstance(worldId, instanceId);
+    return InstancesApi.getInstance(worldId, instanceId).then((resp) => {
         return resp.data;
-    } catch (e: any) {
-        logger.fatal("Error while trying to get instance data: ");
-        //console.error(e);
-        console.log(e.response)
-    }
-    
+    }).catch((e) => {
+        throw new Error(`Error while trying to get instance data: ${e.response}`);
+    })
 }
 
-async function findUserAvatar (userId: string, getOnlyAvatarName?: boolean): Promise<AxiosResponse | undefined | VRSpaceVRCUserAvatar> {
+/**
+ * Fetches user avatar data for the given user ID.
+ *
+ * @param {string} userId - The ID of the user.
+ * @param {boolean} [getOnlyAvatarName] - Optional flag to specify whether to only retrieve the avatar name. Defaults to false.
+ * @returns {Promise<AxiosResponse|VRSpaceVRCUserAvatar>} - The user avatar data.
+ */
+async function findUserAvatar (userId: string, getOnlyAvatarName?: boolean): Promise<AxiosResponse | VRSpaceVRCUserAvatar> {
     logger.write("Getting avatar data for: "+userId)
-    const userData = await getUserInfo(userId).catch(e => {
-        console.log(e);
-        console.log(e.response)
-    });
-    const avatarPictureUrl = userData?.currentAvatarImageUrl;
-    let avatarFileUrl = ""
-    for(let i = 1; i <= 6 ; i++) {
+    return getUserInfo(userId).then(async (userData) => {
+        const avatarPictureUrl = userData?.currentAvatarImageUrl;
+        let avatarFileUrl = ""
+        for(let i = 1; i <= 6 ; i++) {
             if(avatarPictureUrl?.split("/")[i] !== "")
                 avatarFileUrl += avatarPictureUrl?.split("/")[i] + "/";
-    }
-    avatarFileUrl = "https://"+avatarFileUrl;
-    const res = await axios.get(avatarFileUrl, {
-        headers: {
-            'User-Agent': USER_AGENT
-        },
-    }).catch(e => {
-        return undefined;
-    });
-    if (!getOnlyAvatarName) {
-        return res?.data;
-    } else {
-        let avatarName = res?.data.name;
-        avatarName = avatarName.split("Avatar - ")[1].split(" - Image - ")[0];
-        let avatarRes: VRSpaceVRCUserAvatar = {
-            fileId: res?.data.id,
-            ownerId: res?.data.ownerId,
-            avatarName: avatarName,
-            avatarImageUrl: avatarFileUrl,
-            vrcData: res?.data
         }
-        return avatarRes;
-    }
+        avatarFileUrl = "https://"+avatarFileUrl;
+        return axios.get(avatarFileUrl, {
+            headers: {
+                'User-Agent': USER_AGENT
+            },
+        }).then(async (res) => {
+            if (!getOnlyAvatarName) {
+                return res?.data;
+            } else {
+                let avatarName = res?.data.name;
+                avatarName = avatarName.split("Avatar - ")[1].split(" - Image - ")[0];
+                return {
+                    fileId: res?.data.id,
+                    ownerId: res?.data.ownerId,
+                    avatarName: avatarName,
+                    avatarImageUrl: avatarFileUrl,
+                    vrcData: res?.data
+                };
+            }
+        });
+    })
+}
+
+// Right now we don't use this, but it's here for future reference, in case we need to manually parse the cookies
+/**
+ * Extracts cookie data from a given raw cookie string.
+ * @param {string} rawCookie - The raw cookie string to extract data from.
+ * @returns {VRChatCookieFormat} - An object containing the extracted cookie data.
+ */
+function extractCookie(rawCookie: string[1]): VRChatCookieFormat {
+    console.log("raw cookie")
+    console.log(rawCookie)
+    let rawDate = Date.parse(rawCookie.split("Expires=")[1].split(";")[0]);
+    let finalDate = new Date(rawDate);
+    return {
+        key: rawCookie.split("=")[0],
+        value: rawCookie.split("=")[1].split(";")[0],
+        maxAge: Number(rawCookie.split("Max-Age=")[1].split(";")[0]),
+        path: rawCookie.split("Path=")[1].split(";")[0],
+        expires: finalDate,
+        httpOnly: rawCookie.includes("HttpOnly"),
+        sameSite: rawCookie.split("SameSite=")[1].split(";")[0],
+        hostOnly: true,
+        creation: new Date()
+    };
+
 }
 
 
-async function getMyself(): Promise<vrchat.CurrentUser | undefined> {
+async function getMyself(): Promise<vrchat.CurrentUser> {
     logger.write("Getting current user info")
-    try {
-        const resp = await AuthenticationApi.getCurrentUser();
+    return AuthenticationApi.getCurrentUser().then((resp) => {
         return resp.data;
-    } catch (e) {
-        console.error(e);
-    }
+    })
 }
 
-export 
+export
 {
     getAuthCookie,
     getUserInfo,
